Restructuring work on Receive CC
