+++ src/main/java/dk/dtu/matador/objects/chancecards/pay/NewTyres.java
@@ -7,23 +7,23 @@ import dk.dtu.matador.managers.PlayerManager;
 
 import java.util.UUID;
 
-public class buyBeer extends PayCC{
-    private double beerprice = 200;
+public class NewTyres extends PayCC{
+    private double tyrescost = 1000;
 
-    public buyBeer() {
-        super("buyBeer");
+    public NewTyres() {
+        super("newTyres");
     }
-    public buyBeer(double beerprice) {
-        super("buyBeer");
-        this.beerprice = beerprice;
+    public NewTyres(double tyrescost) {
+        super("newTyres");
+        this.tyrescost = tyrescost;
     }
 
     public void doCardAction(UUID playerID) {
-        double money = beerprice;
-        if (PlayerManager.getInstance().getPlayer(playerID).getBalance() > beerprice) {
+        double money = tyrescost;
+        if (PlayerManager.getInstance().getPlayer(playerID).getBalance() > money) {
             PlayerManager.getInstance().getPlayer(playerID).withdraw(money);
 
-            GUIManager.getInstance().showMessage(LanguageManager.getInstance().getString("buyBeer_chancecard_message"));
+            GUIManager.getInstance().showMessage(LanguageManager.getInstance().getString("newTyres_chancecard_message"));
         }
         else {
             GameManager.getInstance().finishGame();
