+++ src/main/java/dk/dtu/matador/objects/chancecards/pay/OilPrices.java
@@ -1,6 +1,6 @@
 package dk.dtu.matador.objects.chancecards.pay;
 
-public class Oilprices extends PayCC{
+public class OilPrices extends PayCC{
     private double oilpricehouse = 500;
     private double oilpricehotel = 2000;
 
@@ -10,7 +10,7 @@ public class Oilprices extends PayCC{
      * @param cardName The programmable card name,
      *                 also used in the language file.
      */
-    Oilprices(String cardName) {
+    OilPrices(String cardName) {
         super(cardName);
     }
 }
