+++ src/main/java/dk/dtu/matador/objects/chancecards/pay/ParkingTicket.java
@@ -7,23 +7,23 @@ import dk.dtu.matador.managers.PlayerManager;
 
 import java.util.UUID;
 
-public class newTyres extends PayCC{
-    private double tyrescost = 1000;
+public class ParkingTicket extends PayCC{
+    private double parkticketprice = 200;
 
-    public newTyres() {
-        super("newTyres");
+    public ParkingTicket() {
+        super("parkTicket");
     }
-    public newTyres(double tyrescost) {
-        super("newTyres");
-        this.tyrescost = tyrescost;
+    public ParkingTicket(double parkticketprice) {
+        super("parkTicket");
+        this.parkticketprice = parkticketprice;
     }
 
     public void doCardAction(UUID playerID) {
-        double money = tyrescost;
+        double money = parkticketprice;
         if (PlayerManager.getInstance().getPlayer(playerID).getBalance() > money) {
             PlayerManager.getInstance().getPlayer(playerID).withdraw(money);
 
-            GUIManager.getInstance().showMessage(LanguageManager.getInstance().getString("newTyres_chancecard_message"));
+            GUIManager.getInstance().showMessage(LanguageManager.getInstance().getString("parkTicket_chancecard_message"));
         }
         else {
             GameManager.getInstance().finishGame();
