+++ src/main/java/dk/dtu/matador/objects/chancecards/pay/PropertyTax.java
@@ -1,6 +1,6 @@
 package dk.dtu.matador.objects.chancecards.pay;
 
-public class Propertytax extends PayCC{
+public class PropertyTax extends PayCC{
     private double housetax = 500;
     private double hoteltax = 2300;
