+++ src/main/java/dk/dtu/matador/objects/chancecards/receive/Aktie.java
@@ -8,19 +8,7 @@ import dk.dtu.matador.managers.PlayerManager;
 import java.util.UUID;
 
 public class Aktie extends ReceiveCC {
-    private double aktieReceiveAmount = 1000.0;
-
     public Aktie() {
-        super("Aktie");
-    }
-    public Aktie(double aktieReceiveAmount) {
-        super("Aktie");
-        this.aktieReceiveAmount = aktieReceiveAmount;
-    }
-
-    public void doCardAction(UUID playerID) {
-        double money = aktieReceiveAmount;
-
-        PlayerManager.getInstance().getPlayer(playerID).deposit(money);
+        super("Aktie", 1000.0);
     }
 }
