+++ src/main/java/dk/dtu/matador/objects/chancecards/receive/FamilieyParty.java
@@ -1,15 +1,7 @@
 package dk.dtu.matador.objects.chancecards.receive;
 
 public class FamilieyParty extends ReceiveCC {
-    private double familleyPartyReceiveAmountFromEveryPlayer = 500.0;
-
-    /**
-     * Initiates a new ChanceCard.
-     *
-     * @param cardName The programmable card name,
-     *                 also used in the language file.
-     */
-    FamilieyParty(String cardName) {
-        super(cardName);
+    public FamilieyParty() {
+        super("FamilieyParty", 500.0);
     }
 }
