+++ src/main/java/dk/dtu/matador/objects/chancecards/receive/OldFurniture.java
@@ -5,19 +5,7 @@ import dk.dtu.matador.managers.PlayerManager;
 import java.util.UUID;
 
 public class OldFurniture extends ReceiveCC {
-    private double furnitureReceiveAmount = 1000.0;
-
     public OldFurniture() {
-        super("OldFurniture");
-    }
-    public OldFurniture(double furnitureReceiveAmount) {
-        super("OldFurniture");
-        this.furnitureReceiveAmount = furnitureReceiveAmount;
-    }
-
-    public void doCardAction(UUID playerID) {
-        double money = furnitureReceiveAmount;
-
-        PlayerManager.getInstance().getPlayer(playerID).deposit(money);
+        super("OldFurniture", 1000.0);
     }
 }
