+++ src/main/java/dk/dtu/matador/objects/chancecards/receive/PremiumBond.java
@@ -5,19 +5,7 @@ import dk.dtu.matador.managers.PlayerManager;
 import java.util.UUID;
 
 public class PremiumBond extends ReceiveCC {
-    private double bondReceiveAmount = 1000.0;
-
     public PremiumBond() {
-        super("PremiumBond");
-    }
-    public PremiumBond(double bondReceiveAmount) {
-        super("PremiumBond");
-        this.bondReceiveAmount = bondReceiveAmount;
-    }
-
-    public void doCardAction(UUID playerID) {
-        double money = bondReceiveAmount;
-
-        PlayerManager.getInstance().getPlayer(playerID).deposit(money);
+        super("PremiumBond", 1000.0);
     }
 }
