+++ src/main/java/dk/dtu/matador/objects/chancecards/receive/Raise.java
@@ -5,19 +5,7 @@ import dk.dtu.matador.managers.PlayerManager;
 import java.util.UUID;
 
 public class Raise extends ReceiveCC {
-    private double raiseReceiveAmount = 1000.0;
-
     public Raise() {
-        super("Raise");
-    }
-    public Raise(double raiseReceiveAmount) {
-        super("Raise");
-        this.raiseReceiveAmount = raiseReceiveAmount;
-    }
-
-    public void doCardAction(UUID playerID) {
-        double money = raiseReceiveAmount;
-
-        PlayerManager.getInstance().getPlayer(playerID).deposit(money);
+        super("Raise", 1000.0);
     }
 }
