+++ src/main/java/dk/dtu/matador/objects/chancecards/receive/TheLocalAuthority.java
@@ -5,19 +5,7 @@ import dk.dtu.matador.managers.PlayerManager;
 import java.util.UUID;
 
 public class TheLocalAuthority extends ReceiveCC {
-    private double localAuthorityReceiveAmount = 3000.0;
-
     public TheLocalAuthority() {
-        super("TheLocalAuthority");
-    }
-    public TheLocalAuthority(double localAuthorityReceiveAmount) {
-        super("TheLocalAuthority");
-        this.localAuthorityReceiveAmount = localAuthorityReceiveAmount;
-    }
-
-    public void doCardAction(UUID playerID) {
-        double money = localAuthorityReceiveAmount;
-
-        PlayerManager.getInstance().getPlayer(playerID).deposit(money);
+        super("TheLocalAuthority", 3000.0);
     }
 }
